Revert "made some comments and implemented in every method from the class car…"

+++ Classes/Cart.php
@@ -1,68 +1,69 @@
-
 <?php
-require('Database.php');
-
 /**
  * Description of Cart
  *
  * @author petersen
  * @author lehmann
  * @author jessen
- * @author Leon Vagnoni
  */
 class Cart {
 
- public function __construct() {
+public $cartId;
+public $devices; //Array of devices which are included in the cart
+public $name;   //name of the cart
+
+public function __construct() {
+
+}
+
     /**
      * @param $devices
      * @param $name
-     * adds a devices to the database 
      */
 public function addCart($devices, $name) {
+    $db = new mysqli('localhost', 'root', '', 'it-ausleihe');
+    print_r($db->connect_error);
+
+    if($db->connect_errno) {
+        die('Es gibt ein Problem mit der Datenbank');
+    }
 
 
     for($i=0; $i<$devices;$i++) {
         $insert = "INSERT INTO 'cart' ('cartID', 'deviceID', 'name') VALUES (". $cartID .", ". $devices[$i] .", ". $name .")";
-        $params = array( 
-            "devices" => $devices,
-            "name" => $name,
-        );
-        $exporter = Database::getDbConnectiont();
-        $exporter->executeQuery($insert, $params);
+        $db->query($insert);
     }
-
 
     }
 
     /**
      * Check the cart about completeness
      * @param $cartID
-     * searchs the typed cartID from the database
      */
 public function checkTotality($cartID) {
+    $db = new mysqli('localhost', 'root', '', 'it-ausleihe');
+    print_r($db->connect_error);
 
+    if($db->connect_errno) {
+        die('Es gibt ein Problem mit der Datenbank');
+    }
     $select = "SELECT 'cartID', 'deviceID' FROM 'cart' WHERE 'cartID' = ". $cartID;
-    $params = array( 
-        "cardID" => $cartID,
-    );
-    $exporter = Database::getDbConnectiont();
-    $exporter->executeQuery($select, $params);
+    $db->query($select);
 }
 
     /**
      * Deletes cart from database
      * @param $cartID
-     * delete the typed cartID from the database
      */
 public function deleteCart($cartID){
- 
-    $params = array(
-        "cardID" => $cartID,
-    );
+    $db = new mysqli('localhost', 'root', '', 'it-ausleihe');
+    print_r($db->connect_error);
 
+    if($db->connect_errno) {
+        die('Es gibt ein Problem mit der Datenbank');
+    }
     $delete = "DELETE FROM 'cart' WHERE 'cartID' = ". $cartID;
-    $exporter = Database::getDbConnectiont();
-    $exporter->executeQuery($delete, $params);
+    $db->query($delete);
 
 }
 
@@ -70,9 +71,15 @@ public function deleteCart($cartID){
      * Modifies the database entry of the cart
      * @param $device
      * @param $name
-     * modifys the typed mType,device and name from the database
      */
 public function modifyCart($mType, $device, $name){ //Modifies the database entry of the cart
+    $db = new mysqli('localhost', 'root', '', 'it-ausleihe');
+    print_r($db->connect_error);
+
+    if($db->connect_errno) {
+        die('Es gibt ein Problem mit der Datenbank');
+    }
+
     if($mType == "add") {
         $modify = "INSERT INTO 'cart' ('cartID', 'deviceID', 'name') VALUES ( ". $cartID .", ". $devices .", ". $name .")";
     }
@@ -80,13 +87,8 @@ public function modifyCart($mType, $device, $name){ //Modifies the database entr
         $modify = "DELETE FROM 'cart' WHERE 'devieID' = ". $deviceID;
     }
     $db->query($modify);
-    $params = array( "mType" => $mType,
-        "device" => $device,
-        "name" => $name,
-    );
-
-    $exporter = Database::getDbConnectiont();
-    $exporter->executeQuery($modify, $params);
 }
 
 }
+
+?>
